Pdf editor done and re-render almost done

+++ AbsolutReader/src/js/renderer.js
@@ -2,25 +2,19 @@ import React, {Component} from 'react';
 import Pdf from 'react-native-pdf'; //Rendering
 import { StyleSheet, View , Dimensions, SafeAreaView, useWindowDimensions} from 'react-native';
 import { Layout, Text, TopNavigation, TopNavigationAction } from '@ui-kitten/components';
-import { encodeToBase64, PDFDocument } from "pdf-lib"; //Adding links
-import {pdf_editor} from './pdf_tools/pdf_editor';
+import {make_links, pdf_editor} from './pdf_tools/pdf_editor';
 
 var RNFS = require('react-native-fs');
 var base64js = require('base64-js');
 
 export default class App extends Component {
 
-	state = {
-
-	  }
-	
 	constructor(props) {
 		super(props);
 
 		this.path = RNFS.DocumentDirectoryPath; //Main path of app
-		this.source = {uri:'http://www.africau.edu/images/default/sample.pdf',cache:true};
 
-		this.state = {chaimager: '{}', chaimager_loaded: 'False'}
+		this.state = {chaimager: '{}', chaimager_loaded: 'False', source:{uri:'http://www.africau.edu/images/default/sample.pdf',cache:true}}
 	}
 
 	chaimager (link) {
@@ -45,7 +39,7 @@ export default class App extends Component {
 	
 	}
 
-	load_chaimager (filepath) {
+	async load_chaimager (filepath) {
 		//Checks if there is a chaimager file and, if not, creates one.
 		//Updates chaimager inside of this.state with the values that got
 		//and updates the PDF
@@ -58,43 +52,45 @@ export default class App extends Component {
 		const chaimager_file_path = this.path + '/chaimager_files/' + chaimager_file_name; //TODO #1
 
 		//Loads json if file exists
-		RNFS.readFile(chaimager_file_path, 'utf8')
-		.then((json) => {
+		json = await RNFS.readFile(chaimager_file_path, 'utf8');
+		
+		if (json) { //Chaimager file exists
+
 			var chaimager_json = JSON.parse(json);
-			//Updating state if file exists
-			this.setState((state) => { return {chaimager: chaimager_json}});},
-			() => {
-				//File does not exists. Create a new one:
-				RNFS.writeFile(chaimager_file_path, '{}', 'utf8');
-				this.setState((state) => {return {chaimager: JSON.parse('{}')};});
-			})
-			.then(() => {
-				//Now that we have the Chaimager json, we need to edit the PDF
+			this.setState((state) => {return {chaimager: chaimager_json}})
+		} 
+		else { //File do not exists
 
-				//Reading the pdf from filepath:
-				RNFS.readFile(filepath, 'base64').then((data) => {
-					return base64js.toByteArray(data);})
-					//Returns PDF as ByteArray
+			RNFS.writeFile(chaimager_file_path, '{}', 'utf8');
+			this.setState((state) => {return {chaimager: JSON.parse('{}')};});
+		}
 	
-				.then((PdfDoc) => {
-					//We now add the links based on the chaimager file
+		//Now that we have the Chaimager json, we need to edit the PDF
+		//Reading the pdf again from filepath:
 
-					//4TESTING:
-					this.setState((state) => {return {
-						chaimager: JSON.parse('{"ids": [{"name":"Moriarty","image":"feioisfe"}, {"name":"Michael","image":"feioisfe"}] }')};});
-					
-					pdf_editor(PdfDoc, 'more');
-					
-					for (var i = 0; i < this.state.chaimager["ids"].length; i++) {
-						//For each id in the Chaimager file we update the pdf:
+		var base64_pdf = await RNFS.readFile(filepath, 'base64');
 
-					};
-
-				}) 
+		var PdfDoc = base64js.toByteArray(base64_pdf); //PDF as byte array so pdf.js library can understand it
 				
+		//4TESTING:
+		this.setState((state) => {return {
+			chaimager: JSON.parse('{"ids": [{"name":"Moriarty","image":"feioisfe"}, {"name":"Michael","image":"feioisfe"}] }')};});
 				
-			})
+		for (var i = 0; i < this.state.chaimager["ids"].length; i++) {
+			
+			//Calling our function to add the links
+			base64_pdf = await pdf_editor(PdfDoc, this.state.chaimager["ids"][i]["name"]);
 
+			//Now we change the source for this. Note that the edited file is not stored on the local filesystem,
+			//rather in the RAM, so the original PDF stays unedited, which is good
+
+			var new_source = {uri:'data:application/pdf;base64,' + base64_pdf + "'"};
+
+			this.setState((state) => {return {
+				source: new_source
+			}});
+
+		}
 	}
 
 	render(){
@@ -103,7 +99,7 @@ export default class App extends Component {
 
 		<View style={styles.container} > 
 			<Pdf
-				source={this.source}
+				source={this.state.source}
 				onLoadComplete={(numberOfPages,filePath)=>{
 					this.load_chaimager(filePath);
 				}}
